@@ -23,10 +23,10 @@
   },
   "dependencies": {
     "classnames": "2.2.5",
-    "clipboard": "2.0.0",
-    "codemirror": "5.37.0",
+    "clipboard": "2.0.1",
+    "codemirror": "5.38.0",
     "color": "3.0.0",
-    "core-js": "2.5.5",
+    "core-js": "2.5.6",
     "file-saver": "1.3.8",
     "golden-layout": "1.5.9",
     "inversify": "4.13.0",
@@ -38,14 +38,14 @@
     "markdown-it": "8.4.1",
     "moment": "2.22.1",
     "po2json": "0.4.5",
-    "react": "16.3.2",
-    "react-dom": "16.3.2",
+    "react": "16.4.0",
+    "react-dom": "16.4.0",
     "react-router": "4.2.0",
     "react-router-dom": "4.2.2",
     "reflect-metadata": "0.1.12",
     "rxjs": "5.5.10",
     "spark-md5": "3.0.0",
-    "tslib": "1.9.0"
+    "tslib": "1.9.1"
   },
   "devDependencies": {
     "@types/chai": "4.1.3",
@@ -57,19 +57,19 @@
     "@types/core-js": "0.9.46",
     "@types/enzyme": "3.1.10",
     "@types/enzyme-adapter-react-16": "1.0.2",
-    "@types/esprima": "4.0.1",
+    "@types/esprima": "4.0.2",
     "@types/file-saver": "1.3.0",
     "@types/glob": "5.0.35",
     "@types/history": "4.6.2",
     "@types/isomorphic-fetch": "0.0.34",
-    "@types/jquery": "3.3.1",
+    "@types/jquery": "3.3.2",
     "@types/keymaster": "1.6.28",
     "@types/markdown-it": "0.0.4",
     "@types/mocha": "5.2.0",
-    "@types/node": "9.6.6",
-    "@types/react": "16.3.12",
+    "@types/node": "10.1.2",
+    "@types/react": "16.3.14",
     "@types/react-dom": "16.0.5",
-    "@types/react-router": "4.0.23",
+    "@types/react-router": "4.0.25",
     "@types/react-router-dom": "4.2.6",
     "@types/react-test-renderer": "16.0.1",
     "archiver": "2.1.1",
@@ -83,7 +83,7 @@
     "esprima": "4.0.0",
     "file-loader": "1.1.11",
     "fork-ts-checker-webpack-plugin": "0.4.1",
-    "git-revision-webpack-plugin": "3.0.0",
+    "git-revision-webpack-plugin": "3.0.3",
     "glob": "7.1.2",
     "handlebars": "4.0.11",
     "handlebars-loader": "1.7.0",
@@ -96,31 +96,31 @@
     "karma": "2.0.2",
     "karma-bamboo-reporter": "0.1.2",
     "karma-chrome-launcher": "2.2.0",
-    "karma-coverage": "1.1.1",
+    "karma-coverage": "1.1.2",
     "karma-firefox-launcher": "1.1.0",
     "karma-mocha": "1.3.0",
     "karma-mocha-reporter": "2.2.5",
     "karma-sourcemap-loader": "0.3.7",
     "karma-webpack": "3.0.0",
-    "less": "3.0.2",
+    "less": "3.0.4",
     "less-loader": "4.1.0",
-    "lint-staged": "7.0.4",
+    "lint-staged": "7.1.2",
     "mini-css-extract-plugin": "0.4.0",
-    "mocha": "5.1.1",
+    "mocha": "5.2.0",
     "pofile": "1.0.10",
     "prettier": "1.12.1",
     "raw-loader": "0.5.1",
-    "react-test-renderer": "16.3.2",
-    "remap-istanbul": "0.11.0",
+    "react-test-renderer": "16.4.0",
+    "remap-istanbul": "0.11.1",
     "style-loader": "0.21.0",
-    "ts-loader": "4.2.0",
-    "tslint": "5.9.1",
+    "ts-loader": "4.3.0",
+    "tslint": "5.10.0",
     "typescript": "2.8.3",
     "url-loader": "1.0.1",
-    "webpack": "4.6.0",
-    "webpack-bundle-analyzer": "2.11.1",
-    "webpack-cli": "2.0.15",
-    "webpack-dev-server": "3.1.3",
+    "webpack": "4.8.3",
+    "webpack-bundle-analyzer": "2.13.1",
+    "webpack-cli": "2.1.4",
+    "webpack-dev-server": "3.1.4",
     "webpack-notifier": "1.6.0"
   },
   "lint-staged": {
