@@ -1,4 +1,6 @@
-import { Observable, Subject } from "rxjs/Rx";
+import { Subject } from "rxjs/Subject";
+import { Observable } from "rxjs/Observable";
+
 import { UserInfo, AuthProvider } from "app/services/login/user-info";
 import { injectable } from "app/di";
 
@@ -20,14 +22,8 @@ export const NotLoggenInStatus: LoginStatus = Object.freeze({
   }
 });
 
-export abstract class ICurrentUserService {
-  abstract getLoginStatus(): LoginStatus;
-  abstract setLoginStatus(loginStatus: LoginStatus): void;
-  abstract loginStatusObservable: Observable<LoginStatus>;
-}
-
 @injectable()
-export class CurrentUserService implements ICurrentUserService {
+export class CurrentUserService {
   private currentLoginStatus = NotLoggenInStatus;
   private loginStatusSubject = new Subject<LoginStatus>();
 
@@ -41,6 +37,6 @@ export class CurrentUserService implements ICurrentUserService {
   }
 
   get loginStatusObservable(): Observable<LoginStatus> {
-    return this.loginStatusSubject;
+    return this.loginStatusSubject.asObservable();
   }
 }
