@@ -1,9 +1,9 @@
 import { RandomHelper } from "app/utils/random-helper";
-import { IAjaxService } from "app/services/infrastructure/ajax-service";
+import { AjaxService } from "app/services/infrastructure/ajax-service";
 import { AppConfig } from "app/services/config/app-config";
 
 export class AppConfigLoader {
-  constructor(private ajaxService: IAjaxService) {}
+  constructor(private ajaxService: AjaxService) {}
 
   async loadData(): Promise<AppConfig> {
     const config = (await this.ajaxService.ajax(
