@@ -1,30 +1,35 @@
 import { injectable } from "app/di";
-import { IAjaxService } from "app/services/infrastructure/ajax-service";
+import { AjaxService } from "app/services/infrastructure/ajax-service";
 
 type DictionaryLike<V> = { [name: string]: V };
-
-export abstract class ILocalizedContentLoader {
-  abstract getFileContent(relativePath: string): Promise<string>;
-}
 
 /**
  * Loads requested files by ajax from content directory using provided locale
  */
 @injectable()
-export class LocalizedContentLoader implements ILocalizedContentLoader {
+export class LocalizedContentLoader {
   private cache: DictionaryLike<string> = {};
 
-  constructor(private ajax: IAjaxService, private localeId: string) {}
+  constructor(private ajax: AjaxService, private localeId: string) {}
 
   async getFileContent(relativePath: string): Promise<string> {
-    //await stay(2000);
+    try {
+      const content = await this.getContentByLocale(this.localeId, relativePath);
+      return content;
+    } catch (ex) {
+      console.log("Failed to load localized version, falling back to English");
+      return this.getContentByLocale("en", relativePath);
+    }
+  }
+
+  private async getContentByLocale(locale: string, relativePath: string): Promise<string> {
     const resKey = `${this.localeId}:${relativePath}`;
     const fromCache = this.cache[resKey];
     if (fromCache) {
       return fromCache;
     }
     try {
-      const result = await this.ajax.getText(`content/${this.localeId}/${relativePath}`);
+      const result = await this.ajax.getText(`content/${locale}/${relativePath}`);
       this.cache[resKey] = result;
       return result;
     } catch (ex) {
