@@ -1,5 +1,6 @@
+import { Observable } from "rxjs/Observable";
+
 import { LoginStatus } from "app/services/login/current-user.service";
-import { Observable } from "rxjs/Rx";
 
 export interface IAuthService {
   init(): Promise<void>;
