@@ -1,7 +1,8 @@
 import * as React from "react";
 import * as cn from "classnames";
 import * as codemirror from "codemirror";
-import { Observable, Subscription } from "rxjs";
+import { Observable } from "rxjs/Observable";
+import { Subscription } from "rxjs/Subscription";
 
 import { ensure } from "app/utils/syntax-helpers";
 
